+++ private/bufpkg/bufplugin/plugin_key_provider.go
@@ -16,9 +16,11 @@ package bufplugin
 
 import (
 	"context"
+	"fmt"
 	"io/fs"
 
 	"github.com/bufbuild/buf/private/bufpkg/bufparse"
+	"github.com/bufbuild/buf/private/pkg/slicesext"
 )
 
 var (
@@ -41,6 +43,17 @@ type PluginKeyProvider interface {
 	GetPluginKeysForPluginRefs(context.Context, []bufparse.Ref, DigestType) ([]PluginKey, error)
 }
 
+// NewStaticPluginKeyProvider returns a new PluginKeyProvider for a static set of PluginKeys.
+//
+// The set of PluginKeys must be unique by FullName. If there are duplicates,
+// an error will be returned.
+//
+// When resolving Refs, the Ref will be matched to the PluginKey by FullName.
+// If the Ref is not found in the set of provided keys, an fs.ErrNotExist will be returned.
+func NewStaticPluginKeyProvider(pluginKeys []PluginKey) (PluginKeyProvider, error) {
+	return newStaticPluginKeyProvider(pluginKeys)
+}
+
 // *** PRIVATE ***
 
 type nopPluginKeyProvider struct{}
@@ -51,4 +64,50 @@ func (nopPluginKeyProvider) GetPluginKeysForPluginRefs(
 	DigestType,
 ) ([]PluginKey, error) {
 	return nil, fs.ErrNotExist
+}
+
+type staticPluginKeyProvider struct {
+	pluginKeysByFullName map[string]PluginKey
+}
+
+func newStaticPluginKeyProvider(pluginKeys []PluginKey) (*staticPluginKeyProvider, error) {
+	var pluginKeysByFullName map[string]PluginKey
+	if len(pluginKeys) > 0 {
+		var err error
+		pluginKeysByFullName, err = slicesext.ToUniqueValuesMap(pluginKeys, func(pluginKey PluginKey) string {
+			return pluginKey.FullName().String()
+		})
+		if err != nil {
+			return nil, err
+		}
+	}
+	return &staticPluginKeyProvider{
+		pluginKeysByFullName: pluginKeysByFullName,
+	}, nil
+}
+
+func (s staticPluginKeyProvider) GetPluginKeysForPluginRefs(
+	_ context.Context,
+	refs []bufparse.Ref,
+	digestType DigestType,
+) ([]PluginKey, error) {
+	pluginKeys := make([]PluginKey, len(refs))
+	for i, ref := range refs {
+		// Only the FullName is used to match the PluginKey. The Ref is not
+		// validated to match the PluginKey as there is not enough information
+		// to do so.
+		pluginKey, ok := s.pluginKeysByFullName[ref.FullName().String()]
+		if !ok {
+			return nil, fs.ErrNotExist
+		}
+		digest, err := pluginKey.Digest()
+		if err != nil {
+			return nil, err
+		}
+		if digest.Type() != digestType {
+			return nil, fmt.Errorf("expected DigestType %v, got %v", digestType, digest.Type())
+		}
+		pluginKeys[i] = pluginKey
+	}
+	return pluginKeys, nil
 }